Update client module and remove explore-container(old code)
