+++ src/app/client/client-routing.module.ts
@@ -1,0 +1,36 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { ClientPage } from './client.page';
+
+const routes: Routes = [
+  {
+    path: '',
+    component: ClientPage,
+    children: [
+      {
+        path: 'home',
+        loadChildren: () => import('../categories/categories.module').then(m => m.CategoriesPageModule)
+      },
+      {
+        path: 'profile',
+        loadChildren: () => import('../profile/profile.module').then(m => m.ProfilePageModule)
+      },
+      {
+        path: '',
+        redirectTo: '/client/home',
+        pathMatch: 'full'
+      }
+    ]
+  },
+  {
+    path: '',
+    redirectTo: '/client/home',
+    pathMatch: 'full'
+  }
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class ClientPageRoutingModule {}
