+++ src/app/client/client.page.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-client',
+  templateUrl: 'client.page.html',
+  styleUrls: ['client.page.scss']
+})
+export class ClientPage {
+
+  constructor() {}
+
+}
